Move Combine and First tests to correct module
